Fixed logic of how the loop is detected in `ReversibleSegment` and `CompleteSegment`
This fixes a mist behavior of `CompleteSegment` when reaching the genesis block in ForkDB which was causing a `loop detected` error to be logged and was not returning the complete segment.

Also improved the documentation around `CompleteSegment` as well as adding tests to ensure loop detection code works as expected.