Introducing `pbbstream.BlockMeta`
This metadata contains the chain agnostic information we possess for each block. It strictly follows `pbstream.Block` semantics and Protobuf definitions except that the `Payload` fields are removed.

By modelling it like this, we gain the ability to decode chain's full block and extract only minimal information from it. We still have the pay the price for holding in memory the `Payload` but at least once read to `pbbstream.BlockMeta`, we don't need the payload anymore.

Soe `bstream` API methods gain `BlockMeta` knowledge for example to read from a one block file only the block meta information.